GUI/requirements.txt
pywebview
gpiozero
pygame
playsound